Add control scheme (#5)
Sets up all gameplay controls for keyboard and mouse in Input Actions
Rename KeyboardInput.cs to PlayerInput.cs
Add all buttons to PlayerInput.cs
